소스코드 가독성 향상
- 소스코드 가독성 향상

- 주석 추가
