+++ main/java/com/mainweb/API/JsonParseAPI.java
@@ -1,0 +1,8 @@
+package com.mainweb.API;
+
+import org.springframework.stereotype.Component;
+
+@Component
+public class JsonParseAPI {
+
+}
