+++ main/java/com/mainweb/DTO/classificationData.java
@@ -1,4 +1,9 @@
 package com.mainweb.DTO;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import lombok.Setter;
+
 /**
  * 저자: 한승헌
  * 기능: 품종 분류 정보를 저장하는 DTO
@@ -8,27 +13,10 @@ package com.mainweb.DTO;
  * name : 품종명
  * value : 섞인 비율
  * */
+@Getter
+@Setter
+@RequiredArgsConstructor
 public class classificationData {
     private String name;
     private int value;
-
-    public classificationData(String name, int value){
-        this.name = name;
-        this. value = value;
-    }
-
-    public String getName() {
-        return name;
-    }
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public int getValue() {
-        return value;
-    }
-
-    public void setValue(int value) {
-        this.value = value;
-    }
 }
