--- main/java/com/mainweb/Service/JsonParserService.java
@@ -1,11 +1,0 @@
-package com.mainweb.Service;
-
-import com.mainweb.DTO.classificationData;
-import org.springframework.stereotype.Component;
-
-import java.util.List;
-
-@Component
-public class JsonParserService {
-
-}
