+++ main/resources/templates/result.html
@@ -52,7 +52,7 @@
             labels: names,
             datasets:[
                 {
-                    label: '당신의 반려견의 비율!',
+                    label: '비율',
                     data: values,
                     backgroundColor : '#24DB00'
                 }
